Specified "ShouldSupportProcess" class attribute

+++ Solutions/PowerShell.Commands/Commands/Branding/RemoveNavigationNode.cs
@@ -2,10 +2,11 @@
 using OfficeDevPnP.PowerShell.Commands.Enums;
 using System.Management.Automation;
 using OfficeDevPnP.Core.Enums;
+using Resources = OfficeDevPnP.PowerShell.Commands.Properties.Resources;
 
 namespace OfficeDevPnP.PowerShell.Commands
 {
-    [Cmdlet(VerbsCommon.Remove, "SPONavigationNode")]
+    [Cmdlet(VerbsCommon.Remove, "SPONavigationNode", SupportsShouldProcess = true)]
     public class RemoveNavigationNode : SPOWebCmdlet
     {
         [Parameter(Mandatory = true)]
@@ -17,9 +18,15 @@ namespace OfficeDevPnP.PowerShell.Commands
         [Parameter(Mandatory = false)]
         public string Header;
 
+        [Parameter(Mandatory = false)]
+        public SwitchParameter Force;
+
         protected override void ExecuteCmdlet()
         {
-            SelectedWeb.DeleteNavigationNode(Title, Header, Location);
+            if (Force || ShouldContinue(string.Format(Resources.RemoveNavigationNode0, Title), Resources.Confirm))
+            {
+                SelectedWeb.DeleteNavigationNode(Title, Header, Location);
+            }
         }
 
     }
